requirements.txt
python-telegram-bot==20.7
Flask==3.1.2
scikit-learn==1.5.2
joblib==1.4.2
requests==2.32.3
numpy==1.26.4
python-dotenv==1.0.1
